+++ Backend/demo/src/main/java/com/example/demo/models/Attraction.java
@@ -9,7 +9,7 @@ public class Attraction {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     private String name;
     private String country;
@@ -35,8 +35,8 @@ public class Attraction {
         this.priceProxim = priceProxim;
     }
 
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    public int getId() { return id; }
+    public void setId(int id) { this.id = id; }
 
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
