+++ Backend/demo/src/main/java/com/example/demo/repositories/GroupRepository.java
@@ -1,6 +1,7 @@
 package com.example.demo.repositories;
+
 import com.example.demo.models.*;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GroupRepository extends JpaRepository<Group, Long> {
+public interface GroupRepository extends JpaRepository<Group, Integer> {
 }
