+++ Backend/demo/src/main/java/com/example/demo/repositories/ReviewRepository.java
@@ -8,8 +8,9 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ReviewRepository extends JpaRepository<Review, Long> {
+public interface ReviewRepository extends JpaRepository<Review, Integer> {
 
     List<Review> findBySupporter(Supporter supporter);
+
     List<Review> findByRating(int rating);
 }
